+++ front/components/dev/devPanelStyles.ts
@@ -1,0 +1,286 @@
+// All styles for the dev console panel UI.
+
+export const FONT =
+  '-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, monospace';
+
+export const S = {
+  // Docked bottom bar
+  docked: {
+    position: "fixed" as const,
+    bottom: 0,
+    left: 0,
+    right: 0,
+    zIndex: 2147483647,
+    fontFamily: FONT,
+    fontSize: 12,
+    color: "#e0e0e0",
+    userSelect: "none" as const,
+  },
+  dockedBar: {
+    display: "flex" as const,
+    alignItems: "center" as const,
+    justifyContent: "space-between" as const,
+    padding: "4px 12px",
+    background: "#111827",
+    borderTop: "1px solid #333",
+    minHeight: 32,
+  },
+  dockedSection: {
+    display: "flex" as const,
+    alignItems: "center" as const,
+    gap: 6,
+  },
+  dockedSep: {
+    width: 1,
+    height: 14,
+    background: "#333",
+    display: "inline-block" as const,
+    margin: "0 2px",
+  },
+  dockedBtn: (active: boolean) => ({
+    padding: "2px 6px",
+    fontSize: 12,
+    border: "none",
+    borderRadius: 3,
+    cursor: "pointer" as const,
+    background: active ? "#ffffff22" : "transparent",
+    color: active ? "#fff" : "#9ca3af",
+  }),
+  dockedTextBtn: (active: boolean) => ({
+    padding: "2px 8px",
+    fontSize: 11,
+    border: active ? "1px solid #7fdbca" : "1px solid transparent",
+    borderRadius: 4,
+    cursor: "pointer" as const,
+    background: active ? "#7fdbca22" : "transparent",
+    color: active ? "#7fdbca" : "#b0b0b8",
+    fontWeight: 500,
+    display: "inline-flex" as const,
+    alignItems: "center" as const,
+    gap: 4,
+  }),
+  dockedIconBtn: {
+    background: "none",
+    border: "none",
+    color: "#9ca3af",
+    cursor: "pointer" as const,
+    fontSize: 13,
+    padding: "2px 4px",
+    borderRadius: 3,
+    lineHeight: 1,
+  },
+  dockedBadge: {
+    background: "#ff6b6b",
+    color: "#fff",
+    borderRadius: 10,
+    padding: "0 5px",
+    fontSize: 9,
+    fontWeight: 700,
+    lineHeight: "14px",
+  },
+  dockedPanel: {
+    position: "fixed" as const,
+    width: 380,
+    maxHeight: "70vh",
+    background: "#1a1a2e",
+    border: "1px solid #333",
+    borderRadius: "10px 10px 0 0",
+    boxShadow: "0 -8px 32px rgba(0,0,0,0.5)",
+    display: "flex" as const,
+    flexDirection: "column" as const,
+    overflow: "hidden" as const,
+    zIndex: 2147483647,
+  },
+  dockedPanelDragHandle: {
+    display: "flex" as const,
+    alignItems: "center" as const,
+    justifyContent: "center" as const,
+    padding: "2px 0",
+    cursor: "grab" as const,
+    background: "#111827",
+    borderBottom: "1px solid #2a2a4a",
+    flexShrink: 0,
+  },
+  dockedPanelGrip: {
+    fontSize: 12,
+    color: "#555",
+    lineHeight: 1,
+    letterSpacing: 2,
+  },
+
+  // Floating panel
+  floatPanel: {
+    position: "fixed" as const,
+    zIndex: 2147483647,
+    minWidth: 500,
+    maxHeight: "80vh",
+    background: "#1a1a2e",
+    border: "1px solid #333",
+    borderRadius: 10,
+    boxShadow: "0 8px 32px rgba(0,0,0,0.5)",
+    fontFamily: FONT,
+    fontSize: 12,
+    color: "#e0e0e0",
+    display: "flex" as const,
+    flexDirection: "column" as const,
+    userSelect: "none" as const,
+    overflow: "hidden" as const,
+  },
+  floatHeader: {
+    display: "flex" as const,
+    alignItems: "center" as const,
+    justifyContent: "space-between" as const,
+    padding: "6px 12px",
+    background: "#111827",
+    cursor: "grab" as const,
+    borderBottom: "1px solid #2a2a4a",
+    flexShrink: 0,
+  },
+
+  // Shared panel internals
+  panelHeader: {
+    display: "flex" as const,
+    alignItems: "center" as const,
+    justifyContent: "space-between" as const,
+    padding: "8px 12px",
+    background: "#16213e",
+    borderBottom: "1px solid #333",
+    flexShrink: 0,
+  },
+  headerTitle: {
+    fontWeight: 700,
+    fontSize: 13,
+    color: "#7fdbca",
+  },
+  headerBtn: {
+    background: "none",
+    border: "none",
+    color: "#a1a1aa",
+    cursor: "pointer" as const,
+    fontSize: 16,
+    padding: "2px 6px",
+    borderRadius: 4,
+    lineHeight: 1,
+  },
+  toolbar: {
+    padding: "8px 12px",
+    borderBottom: "1px solid #333",
+    display: "flex" as const,
+    flexDirection: "column" as const,
+    gap: 6,
+    flexShrink: 0,
+  },
+  toolbarRow: {
+    display: "flex" as const,
+    alignItems: "center" as const,
+    justifyContent: "space-between" as const,
+  },
+  toolbarLabel: {
+    fontSize: 11,
+    color: "#b0b0b8",
+  },
+  btnGroup: {
+    display: "flex" as const,
+    gap: 2,
+  },
+  toolbarBtn: (active: boolean) => ({
+    padding: "3px 10px",
+    fontSize: 10,
+    fontWeight: active ? 700 : 400,
+    border: active ? "1px solid #7fdbca" : "1px solid #555",
+    borderRadius: 4,
+    cursor: "pointer" as const,
+    background: active ? "#7fdbca88" : "transparent",
+    color: active ? "#fff" : "#9ca3af",
+  }),
+  search: {
+    padding: "8px 12px",
+    borderBottom: "1px solid #2a2a4a",
+    flexShrink: 0,
+  },
+  searchInput: {
+    width: "100%",
+    background: "#0f0f23",
+    border: "1px solid #555",
+    borderRadius: 6,
+    padding: "6px 10px",
+    color: "#e0e0e0",
+    fontSize: 12,
+    outline: "none",
+  },
+  list: {
+    overflowY: "auto" as const,
+    flex: 1,
+    padding: "4px 0",
+  },
+  row: {
+    display: "flex" as const,
+    alignItems: "center" as const,
+    padding: "5px 12px",
+    gap: 8,
+    borderBottom: "1px solid #1a1a3a",
+  },
+  flagName: {
+    flex: 1,
+    fontSize: 11,
+    wordBreak: "break-all" as const,
+    lineHeight: 1.3,
+  },
+  flagDesc: {
+    fontSize: 10,
+    color: "#a1a1aa",
+    marginTop: 2,
+  },
+  triToggle: {
+    display: "flex" as const,
+    gap: 2,
+    flexShrink: 0,
+  },
+  toggleBtn: (active: boolean, color: string) => ({
+    padding: "2px 8px",
+    fontSize: 10,
+    fontWeight: active ? 700 : 400,
+    border: active ? `1px solid ${color}` : "1px solid #555",
+    borderRadius: 4,
+    cursor: "pointer" as const,
+    background: active ? `${color}88` : "transparent",
+    color: active ? "#fff" : "#9ca3af",
+  }),
+  badge: (active: boolean) => ({
+    display: "inline-block" as const,
+    width: 6,
+    height: 6,
+    borderRadius: "50%",
+    background: active ? "#7fdbca" : "transparent",
+    marginRight: 4,
+    flexShrink: 0,
+  }),
+  sectionLabel: {
+    padding: "6px 12px 4px",
+    fontSize: 10,
+    fontWeight: 700,
+    textTransform: "uppercase" as const,
+    letterSpacing: 0.5,
+    color: "#9ca3af",
+    borderBottom: "1px solid #1a1a3a",
+  },
+  footer: {
+    padding: "6px 12px",
+    borderTop: "1px solid #333",
+    display: "flex" as const,
+    justifyContent: "space-between" as const,
+    alignItems: "center" as const,
+    fontSize: 10,
+    color: "#9ca3af",
+    flexShrink: 0,
+  },
+  resetBtn: {
+    background: "none",
+    border: "1px solid #555",
+    color: "#ff6b6b",
+    cursor: "pointer" as const,
+    fontSize: 10,
+    padding: "2px 8px",
+    borderRadius: 4,
+  },
+} as const;
